requirements.txt
fastapi
uvicorn
paddlepaddle
paddleocr
pillow